third_party/WebKit/LayoutTests/webaudio/dom-exceptions-expected.txt

Tests DOM exception messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS context.createBuffer(99, 1, context.sampleRate) threw exception SyntaxError: Failed to construct 'AudioBuffer': requested number of channels (99) exceeds maximum (32).
PASS context.createBuffer(1, 1, 1) threw exception SyntaxError: Failed to construct 'AudioBuffer': requested sample rate (1) does not lie in the allowed range of 22050-96000 Hz.
PASS context.createBuffer(1, 1, 1e6) threw exception SyntaxError: Failed to construct 'AudioBuffer': requested sample rate (1.00000e+6) does not lie in the allowed range of 22050-96000 Hz.
PASS context.createBuffer(new ArrayBuffer(100), true) threw exception SyntaxError: Failed to construct 'AudioBuffer': invalid audio data in ArrayBuffer..
PASS context.decodeAudioData(null, function() {}, function () {}) threw exception SyntaxError: Failed to execute 'decodeAudioData' on 'AudioContext': invalid ArrayBuffer for audioData..
PASS context.createMediaElementSource(null) threw exception InvalidStateError: Failed to construct 'MediaElementAudioSourceNode': invalid HTMLMedialElement..
PASS context.createMediaStreamSource(null) threw exception InvalidStateError: Failed to construct 'MediaStreamAudioSourceNode': invalid MediaStream source.
PASS context.createScriptProcessor(1, 1, 1) threw exception SyntaxError: Failed to construct 'ScriptProcessorNode': buffer size (1) must be a power of two between 256 and 16384..
PASS context.createScriptProcessor(4096, 100, 1) threw exception SyntaxError: Failed to construct 'ScriptProcessorNode': number of input channels (100) exceeds maximum (32)..
PASS context.createScriptProcessor(4096, 1, 100) threw exception SyntaxError: Failed to construct 'ScriptProcessorNode': number of output channels (1) exceeds maximum (32)..
PASS context.createChannelSplitter(0) threw exception SyntaxError: Failed to construct 'ChannelSplitterNode': number of outputs (0) must be between 1 and 32..
PASS context.createChannelSplitter(99) threw exception SyntaxError: Failed to construct 'ChannelSplitterNode': number of outputs (99) must be between 1 and 32..
PASS context.createChannelMerger(0) threw exception SyntaxError: Failed to construct 'ChannelMergerNode': number of inputs (0) must be between 1 and 32..
PASS context.createChannelMerger(99) threw exception SyntaxError: Failed to construct 'ChannelMergerNode': number of inputs (99) must be between 1 and 32..
PASS context.createPeriodicWave(null, null) threw exception SyntaxError: Failed to construct 'PeriodicWave': invalid real array.
PASS context.createPeriodicWave(new Float32Array(10), null) threw exception SyntaxError: Failed to construct 'PeriodicWave': invalid imaginary array.
PASS context.createPeriodicWave(new Float32Array(10), new Float32Array(7)) threw exception SyntaxError: Failed to construct 'PeriodicWave': length of real array (10) and length of imaginary array (7) must match..
PASS node.fftSize = 42 threw exception NotSupportedError: Failed to set the 'fftSize' property on 'AnalyserNode': FFT size (42) must be a power of two between 32 and 2048, inclusive.
PASS node.getChannelData(2) threw exception SyntaxError: Failed to execute 'getChannelData' on 'AudioBuffer': channel index (2) exceeds number of channels (1).
PASS node.connect(null, 0, 0) threw exception SyntaxError: Failed to execute 'connect' on 'AudioNode': invalid destination node..
PASS node.connect(context.destination, 100, 0) threw exception IndexSizeError: Failed to execute 'connect' on 'AudioNode': output index (100) exceeds number of outputs (1)..
PASS node.connect(context.destination, 0, 100) threw exception IndexSizeError: Failed to execute 'connect' on 'AudioNode': input index (100) exceeds number of inputs (1)..
PASS node.connect(node2.gain, 100) threw exception IndexSizeError: Failed to execute 'connect' on 'AudioNode': output index (100) exceeds number of outputs (1)..
PASS node.disconnect(99) threw exception IndexSizeError: Failed to execute 'disconnect' on 'AudioNode': output index (99) exceeds number of outputs (1)..
PASS node.connect(otherContext.destination) threw exception SyntaxError: Failed to execute 'connect' on 'AudioNode': cannot connect to a destination belonging to a different audio context..
PASS node.channelCount = 99 threw exception InvalidStateError: Failed to set the 'channelCount' property on 'AudioNode': channel count (99) must be between 1 and 32..
PASS node.channelCountMode = mode threw exception InvalidStateError: Failed to set the 'channelCountMode' property on 'AudioNode': invalid mode 'fancy'; must be 'max', 'clamped-max', or 'explicit'..
PASS node.channelInterpretation = mode threw exception InvalidStateError: Failed to set the 'channelInterpretation' property on 'AudioNode': invalid interpretation 'fancy'; must be 'speakers' or 'discrete'..
PASS context.destination.channelCount = 99 threw exception InvalidStateError: Failed to set the 'channelCount' property on 'AudioDestinationNode': channel count (99) must be between 1 and 2..
PASS new webkitOfflineAudioContext(99, 100, context.sampleRate) threw exception SyntaxError: Failed to construct 'OfflineAudioContext': number of channels (99) exceeds maximum (10)..
PASS new webkitOfflineAudioContext(1, 100, 1) threw exception SyntaxError: Failed to construct 'OfflineAudioContext': sample rate (1) must be in the range 44100-96000 Hz..
PASS new webkitOfflineAudioContext(1, 100, 1e6) threw exception SyntaxError: Failed to construct 'OfflineAudioContext': sample rate (1.00000e+6) must be in the range 44100-96000 Hz..
PASS successfullyParsed is true

TEST COMPLETE
